+++ src/main/resources/application.properties
@@ -1,7 +1,11 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://flag-db.cbutsy3iam3a.ap-northeast-2.rds.amazonaws.com:3306/flag_db
 spring.datasource.username=admin
-spring.datasource.password=flagdb1!
-//spring.datasource.password=${DBPW}
+spring.datasource.password=${DBPW}
 
 spring.jpa.hibernate.naming.physical-strategy = org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+
+spring.thymeleaf.cache=false 
+spring.thymeleaf.enabled=true 
+spring.thymeleaf.prefix=classpath:/templates/
+spring.thymeleaf.suffix=.html
