+++ src/test/java/com/flag/flag_back/FlagBackApplicationTests.java
@@ -2,7 +2,7 @@ package com.flag.flag_back;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-@SpringBootTest
+@SpringBootTests
 class FlagBackApplicationTests {
 
 	@Test
